Don't make bar automatically transparent for non-stack bar with raw fields on level of detail channels (color/size/detail).
Basically, this is not always correct in the case that users over-encode bar's color with the same field as the categorical axis.
(It should not use transparent bar in such case.)
I think this case is actually even more common that layered bar so I agree with @domoritz in #2027 that we should not always use opacity.

Fix #2027