Add identifier transform for default point selections.
Fixes #2221. Default point selections are projected over an ID field.
However, the default `_id` is volatile -- if the spec is parsed again,
different IDs may be generated based on global Vega state. As a result,
breaks the ability to snapshot state via view.getState/setState.

With vega/vega-view#10, an identifier transform was introduced which
persists its state in an internal signal. This transform is added
whenever new tuples are derived (i.e., at the source and post-aggregation).