Add mark config properties to mark definition
Rationale: now that config is a top-level property -- there is no direct way to customize each individual single view.

(The only possible way is to use `style`, but that is pretty indirectly and make more sense when one want to apply the same styles to multiple parts.)

An alternative solution would be implementing encoding channels for all of these properties. However, that requires careful thinking about scale and scale ranges + legends for all of the properties we may add.  Given that's a non-trivial amount of design work, I don't think it's gonna happen anytime soon so this seems like the most natural thing to do.