Make EncodingSortField.field required
If `EncodingSortField.field` is not required, then `EncodingSortField` has no required fields at all, meaning it can match empty JSON. Having the JSON "sort": { } be interpreted to mean that the scale should be sorted by that field doesn't seem intuitive.

This PR makes `field` a required property of `EncodingSortField`. This could lead to redundant field specifications in the case where the user wants to sort by the same field as the scale, but otherwise is harmless.

An alternative solution could be to have a discriminator property (e.g. a required "sortType": "field" prop), but it seems easier to just make `field` non-optional.