fix: generate unescaped boxplot calculated field aliases (#9284)
## PR Description

This change ensures all field aliases are unescaped when generated.
Before this change if a field was referenced like `"field": "one\\.two"`
the boxplot code would generate calculated fields like `"as":
"lower_box_one\\.two"`. This breaks downstream in the compiled Vega code
which expects aliases to not include escape sequences.

As far as I could tell there is no issue open for this. The
contributing.md doc explicitly calls out it is not necessary to open
one, but I'm happy to do so if it helps.

---------

Co-authored-by: Dominik Moritz <domoritz@gmail.com>