fix: align stack order with color domain (#9641)
## PR Description

This PR addresses long-standing issues (dating back to 2016) related to
misalignment between the order of stacked marks and the domain in the
color encoding. It potentially closes #1734, #6203, and #9496.

### Proposed Solution
A new method has been added to the unit parser to check if a color
domain is defined in the unit. If a domain is present and no custom
order is set, additional encoding is injected to ensure that the color
order of the mark matches the order in the legend.

### Note
This is my first contribution to the project. If my code doesn't meet
the project's standards or if the process I followed is incorrect, I
apologise. I'm happy to refactor, feedback is very welcome!

### Examples
Below are examples of specs rendered with the current code (left) and
the same specs rendered with the proposed solution (right). All specs
share this color encoding: `"scale": {"domain": ["B", "A", "C"],
"range": ["red", "blue", "green"]}`

<img width="830" height="1440" alt="examples"
src="https://github.com/user-attachments/assets/785c734f-9eee-4e08-9f2a-74b2a9351ce9"
/>